requirements.txt
Pillow>=10.0.0
exifread>=3.0.0
requests>=2.31.0
flask>=3.0.0
anthropic>=0.18.0
cloudinary>=1.36.0
